+++ test/Unit/RuleSet/Php70Test.php
@@ -234,7 +234,7 @@ final class Php70Test extends ExplicitRuleSetTestCase
         'line_ending' => true,
         'linebreak_after_opening_tag' => true,
         'list_syntax' => [
-            'syntax' => 'short',
+            'syntax' => 'long',
         ],
         'logical_operators' => true,
         'lowercase_cast' => true,
